2.0a1 - Commit final

+++ cms-special/admin/updates/index.php
@@ -1,91 +1,4 @@
-<?php
-
-$invalid = false;
-
-if (isset($_COOKIE['ADMIN_TOKEN'])) {
-    if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/tokens/" . $_COOKIE['ADMIN_TOKEN'])) {
-
-    } else {
-        die("<script>location.href = '/cms-special/admin'</script>");
-    }
-} else {
-    die("<script>location.href = '/cms-special/admin'</script>");
-}
-
-if (isset($_POST['password'])) {
-    if (password_verify($_POST['password'], file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/password"))) {
-        die("<script>location.href = '/cms-special/admin/home';</script>");
-        return;
-    } else {
-        $invalid = true;
-    }
-}
-
-$size = 0;
-
-function getData(string $dir, $ignoreUploadDir = false) {
-    global $size;
-    $dircontent = scandir($dir);
-    foreach ($dircontent as $direl) {
-        if ($ignoreUploadDir && ($direl == "/upload" || $dir . "/" . $direl == $_SERVER['DOCUMENT_ROOT'] . "/resources/upload")) {} else {
-            if ($direl == "." || $direl == ".." || $direl == ".git") {} else {
-                if (is_link($dir . "/" . $direl)) {} else {
-                    if (is_dir($dir . "/" . $direl)) {
-                        getData($dir . "/" . $direl);
-                    } else {
-                        $size = $size + filesize($dir . "/" . $direl);
-                    }
-                }
-            }
-        }
-    }
-}
-
-?>
-
-<?php ob_start();echo("<!--\n\n" . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/resources/private/license") . "\n\n-->") ?>
-<?php
-
-if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
-    $ready = true;
-} else {
-    $ready = false;
-}
-
-?>
-
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="/resources/css/admin.css">
-    <link rel="stylesheet" href="/resources/css/fonts-import.css">
-    <link rel="stylesheet" href="/resources/css/ui.css">
-    <title><?php
-    
-    if ($ready) {
-        echo("Mise à jour et sécurité - Administration du site - " . file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/sitename"));
-    } else {
-        echo("Administration du site - MPCMS");
-    }
-
-    ?></title>
-    <?php
-        if (!$ready) {
-            die("<script>location.href = '/cms-special/setup';</script></head>");
-        }
-    ?>
-    <?php include_once $_SERVER['DOCUMENT_ROOT'] . "/resources/private/header.php"; ?>
-</head>
-<body>
-    <div id="settings">
-    <?php $banner = "/resources/image/codename.jpg";if (getAvgLuminance($_SERVER['DOCUMENT_ROOT'] . "/resources/image/codename.jpg") > 50) {$blackBannerText = true;} else {$blackBannerText = false;}function getAvgLuminance($filename, $num_samples=30) {$img = imagecreatefromjpeg($filename);$width = imagesx($img);$height = imagesy($img);$x_step = intval($width/$num_samples);$y_step = intval($height/$num_samples);$total_lum = 0;$sample_no = 1;for ($x=0; $x<$width; $x+=$x_step) {for ($y=0; $y<$height; $y+=$y_step) {$rgb = imagecolorat($img, $x, $y);$r = ($rgb >> 16) & 0xFF;$g = ($rgb >> 8) & 0xFF;$b = $rgb & 0xFF;$lum = ($r+$r+$b+$g+$g+$g)/6;$total_lum += $lum;$sample_no++;}}$avg_lum  = $total_lum / $sample_no;return ($avg_lum / 255) * 100;}getData($_SERVER['DOCUMENT_ROOT']);$sizestr = $size . " octets";if ($size > 1024) {if ($size > 1048576) {if ($size > 1073741824) {$sizestr = round($size / 1073741824, 2) . " Gio";} else {$sizestr = round($size / 1048576, 2) . " Mio";}} else {$sizestr = round($size / 1024, 2) . " Kio";}} else {$sizestr = $size . " octets";}$sizestr = str_replace(".", ",", $sizestr); ?>
-    <div id="banner" style='background-image: url("<?= $banner ?>");'>
-        <center><table style="width:100%;"><tr><td style="width:50%;"><img style="float:right;" id="banner-logo" src="/resources/upload/siteicon.png"><td><td><span style="float:left;" id="adminb" <?php if ($blackBannerText) {echo("class=\"banner-black\"");} ?>><span id="banner-name" <?php if ($blackBannerText) {echo("class=\"banner-black\"");} ?>><?= file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent/sitename") ?><br></span><?= file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/version") ?> <?= file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/codename") ?> • <?= $sizestr ?></span></td></tr></table></center>
-    </div><div id="navigation"><a href="/cms-special/admin/home" class="sblink">Administration</a> &gt; <a href="/cms-special/admin/housekeeping" class="sblink">Mise à jour et sécurité</a></div>
-        <h3>État global</h3>
+<?php $pageConfig = [ "domName" => "Mise à jour et sécurité", "headerName" => "Mise à jour et sécurité" ]; include_once $_SERVER['DOCUMENT_ROOT'] . "/cms-special/admin/\$resources/precontent.php"; ?>
         <?php
 
         $currentVersion = file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/version");
@@ -121,7 +34,7 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
         }
 
         ?>
-    <h3>Statistiques</h3>
+    <h3>Informations</h3>
     <ul><li>
     <?php
 
@@ -172,8 +85,6 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
     getData($_SERVER['DOCUMENT_ROOT'] . "/cms-special");
     $mpcmsSize = $mpcmsSize + $size;
     getData($_SERVER['DOCUMENT_ROOT'] . "/widgets");
-    $mpcmsSize = $mpcmsSize + $size;
-    getData($_SERVER['DOCUMENT_ROOT'] . "/vendor");
     $mpcmsSize = $mpcmsSize + $size;
 
     $size = 0;
@@ -236,9 +147,9 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
     <h4>Version actuelle (<?= file_get_contents($_SERVER['DOCUMENT_ROOT'] . "/api/version") ?>)</h4>
     <?php
     
-    try {
+    if (file_get_contents("https://gitlab.com/minteck-projects/mpcms/changelog/raw/master/changelog/" . str_replace(" ", "%20", $currentVersion))) {
         echo(file_get_contents("https://gitlab.com/minteck-projects/mpcms/changelog/raw/master/changelog/" . str_replace(" ", "%20", $currentVersion)));
-    } catch (Notice $err) {
+    } else {
         echo("<i>Aucune information concernant votre version de Minteck Projects CMS</i>");
     }
     
@@ -252,13 +163,12 @@ if (file_exists($_SERVER['DOCUMENT_ROOT'] . "/data/webcontent")) {
         } else {
             echo("<h4>Version stable en circulation (" . $latestVersion . ")</h4>");
         }
-        try {
+        if (file_get_contents("https://gitlab.com/minteck-projects/mpcms/changelog/raw/master/changelog/" . str_replace(" ", "%20", $latestVersion))) {
             echo(file_get_contents("https://gitlab.com/minteck-projects/mpcms/changelog/raw/master/changelog/" . str_replace(" ", "%20", $latestVersion)));
-        } catch (Notice $err) {
+        } else {
             echo("<i>Aucune information concernant la dernière version de Minteck Projects CMS</i>");
         }
     }
 
     ?>
-</body>
-</html>
+<?php include_once $_SERVER['DOCUMENT_ROOT'] . "/cms-special/admin/\$resources/postcontent.php"; ?>
